requirements.txt
Flask==0.9
Flask-DebugToolbar==0.7.1
Flask-Mail==0.7.4
Flask-Script==0.5.3
Flask-WTF==0.8.2
Jinja2==2.6
PyYAML==3.10
Scrapy==0.16.4
Twisted==12.3.0
WTForms==1.0.2
Werkzeug==0.8.3
argparse==1.2.1
blinker==1.2
distribute==0.6.34
feedparser==5.1.3
flask-mongoengine==0.6
ipython==0.13.1
lxml==3.1beta1
mongoengine==0.7.9
pyOpenSSL==0.13
pymongo==2.4.1
python-dateutil==2.1
six==1.2.0
virtualenv==1.4.9
w3lib==1.2
wsgiref==0.1.2
zope.interface==4.0.3
